+++ app/assets/styles/theme/variables.scss
@@ -39,8 +39,10 @@ $navigation-hover: $blue-light;
 $black              :  #000;
 $white              :  #fff;
 $offWhite           :  #fdfdfd;
-$whiteSmoke         :  #fafafa;
-$whiteSmoke-darker  :  #f5f5f5;
+$whitesmoke         :  #fafafa;
+$whiteSmoke         :    $whitesmoke;
+$whitesmoke-darker  :  #f5f5f5;
+$whiteSmoke-darker  :    $whitesmoke-darker;
 $snowDriftWhite     :  #fafcfa;
 
 // Grays
