+++ app/components/base/property-list/index.ts
@@ -2,3 +2,4 @@ export * from "./property-list.component";
 export * from "./property-list.module";
 export * from "./text-property.component";
 export * from "./table-property.component";
+export * from "./property-group.component";
