+++ app/components/base/property-list/property-group.html
@@ -1,0 +1,17 @@
+<div class="group-header" (click)="toogleCollapsed()">
+    <div class="collapse-icon">
+        <i class="fa fa-caret-right"*ngIf="collapsed"></i>
+        <i class="fa fa-caret-down" *ngIf="!collapsed"></i>
+    </div>
+    <h4 class="noselect">{{label}}</h4>
+    <span *ngIf="collapsed">
+        <ng-content select="[collapsed-preview]"></ng-content>
+    </span>
+</div>
+<div class="group-content" *ngIf="!collapsed">
+    <div class="warning-message" *ngIf="warningMessage">
+        <i class="fa fa-warning"></i>
+        <span>{{warningMessage}}</span>
+    </div>
+    <ng-content></ng-content>
+</div>
