+++ app/components/job/details/job-details.html
@@ -28,9 +28,7 @@
             <md-tab-group>
                 <md-tab>
                     <template md-tab-label>Properties</template>
-                    <div class="tab-padding">
-                        <bex-job-properties [job]="job"></bex-job-properties>
-                    </div>
+                    <bex-job-properties [job]="job"></bex-job-properties>
                 </md-tab>
                 <md-tab>
                     <template md-tab-label>Environment Settings</template>
