+++ app/components/node/details/node-details.html
@@ -20,9 +20,7 @@
             <md-tab-group>
                 <md-tab>
                     <template md-tab-label>Properties</template>
-                    <div class="tab-padding">
-                        <bex-node-properties [node]="node"></bex-node-properties>
-                    </div>
+                    <bex-node-properties [node]="node"></bex-node-properties>
                 </md-tab>
                 <md-tab>
                     <template md-tab-label>Recent tasks</template>
