+++ app/components/pool/details/pool-details.html
@@ -60,9 +60,7 @@
                 </md-tab>
                 <md-tab>
                     <template md-tab-label>Properties</template>
-                    <div class="tab-padding">
-                        <bex-pool-properties [pool]="pool"></bex-pool-properties>
-                    </div>
+                    <bex-pool-properties [pool]="pool"></bex-pool-properties>
                 </md-tab>
             </md-tab-group>
         </md-card>
