+++ app/components/task/details/task-details.html
@@ -30,9 +30,7 @@
                 </md-tab>
                 <md-tab>
                     <template md-tab-label>Properties</template>
-                    <div class="tab-padding">
-                        <bex-task-properties [job]="job" [task]="task"></bex-task-properties>
-                    </div>
+                    <bex-task-properties [job]="job" [task]="task"></bex-task-properties>
                 </md-tab>
                 <md-tab>
                     <template md-tab-label>Environment Settings</template>
